Support all IRQ lines, implement BSOD
git-svn-id: https://subversion.ews.illinois.edu/svn/fa16-ece391/_projects/loliOS@8451 0df279af-beec-47f8-8c4d-85a890efaaa8

+++ student-distrib/idt.h
@@ -105,10 +105,10 @@ irq_handler_t irq_handlers[16];
 void idt_init(void);
 
 /* Registers an IRQ handler. */
-void register_irq_handler(uint8_t irq_num, void (*callback)(void));
+void register_irq_handler(uint32_t irq_num, void (*callback)(void));
 
 /* Unregisters an IRQ handler */
-void unregister_irq_handler(uint8_t irq_num);
+void unregister_irq_handler(uint32_t irq_num);
 
 /* Interrupt handler routine */
 void handle_interrupt(int_regs_t *regs);
